Collapse "tool-search-web"
